Fixed the authorization bug
Restructured the whole package because the token wasn't being distributed to Fetcher.send(). All methods with exception to fetcher.js are now class based and requires initialization with the new Utilities class.

Also migrated all of the query header functions into one method in the utilities function to cover them all. This can now be accessed at Utilities.generateQueryHeaders.

Added one new test: authTest.js. This will test if the authorization is working or not. Check the isFavourite property which should show as true if token is working. Character may need to be changed out depending if the user has favourited Myne or not.